Fix transcript widget
